==> 1-clean/3-events.ipynb <==
# Path

import sys
import os

project_path = r'/home/craxiss/Documents/projects/kapsul_etkinlik_v2'

os.chdir(project_path)
sys.path.append(project_path)

# Import

from helpers import column_formatter, utf8_convert
from charset_normalizer import detect
from config import data_path as path
from datetime import datetime
from json import dump, load
import pandas as pd
import openpyxl

# Read

lesson_all = pd.read_excel(
    path+'/2-cleaned/basvuru.xlsx', sheet_name='Dersler')
roll_called_lessons = pd.read_excel(
    path+'/2-cleaned/yoklama.xlsx', index_col='Unnamed: 0', sheet_name='General Info')

# Merge & Concat

lesson_all.dropna(inplace=True)
lesson_all.index = lesson_all[0]
lesson_all = lesson_all.T
lesson_all = lesson_all.iloc[:1]
lesson_all.index = ['Senkron Adı']
lesson_all=lesson_all.T

roll_called_lessons=roll_called_lessons.merge(lesson_all, right_index=True, left_index=True)

index = list(set(lesson_all.index)-set(roll_called_lessons.index))

empty_df = pd.DataFrame(columns = roll_called_lessons.columns, index=index)
empty_df['Senkron Adı'] = lesson_all.T[index].T
concat_data = pd.concat([roll_called_lessons, empty_df])

lessons = concat_data.sort_values('TARİH')
lessons_format = pd.DataFrame(lessons['Senkron Adı'], index=lessons.index)

lessons['Normal Adı'] = lessons.index
lessons.index = list(range(lessons.shape[0]))
lessons.drop(7, axis=0, inplace=True)

lessons.index = lessons['Normal Adı']
lessons = lessons[lessons.columns[:-2]]
lessons.index.name = None

lessons['EĞİTİMCİ'] = column_formatter(lessons['EĞİTİMCİ'])

lessons['SÜRE'] = (lessons['BİTİŞ SAATİ']-lessons['BAŞLANGIÇ SAATİ'])
lessons.drop('BİTİŞ SAATİ', axis=1, inplace=True)

lessons = lessons.rename(columns={'BAŞLANGIÇ SAATİ': 'SAAT', 'EĞİTİMCİ':'egitimci'})

lessons.columns = list(map(lambda x: x.lower(), list(lessons.columns)))

lessons['eğitim'] = lessons.index
lessons.index = list(range(len(lessons)))

lessons

# Write

with pd.ExcelWriter(path+'/3-result/events.xlsx') as writer:
    lessons.to_excel(writer, sheet_name='Events')
    lessons_format.to_excel(writer, sheet_name='Format')
